refactor: move api io to view
